+++ app/lib/core/trale_notifier.dart
@@ -5,7 +5,6 @@ import 'package:material_ui/material_ui.dart';
 import 'package:quantumphysique/quantumphysique.dart';
 import 'package:trale/core/backup_interval.dart';
 import 'package:trale/core/first_day.dart';
-import 'package:trale/core/health_connect_service.dart';
 import 'package:trale/core/interpolation.dart';
 import 'package:trale/core/measurement_database.dart';
 import 'package:trale/core/measurement_interpolation.dart';